backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
httpx==0.27.2
openai-agents==0.0.7
mcp==1.2.0
pydantic==2.10.6
pydantic-settings==2.6.1
